Fix exception handling not to depend on specific library
It is a problem when starting jetty.
When using jetty's profile, this catalina class does not exist, so an
error occurred and the server can not be started.
Although it may not be the best method, it can be rewritten so that it
is resolved by naming and does not depend on the catalina class.