Addition of Japanese Sort processing
Two processes of wash-change process and change of sort key.

 - Washing process
In the initial process of scanning, there are READ analyzed by the
system and SORT read from the tag.
Process of deciding what to use as a key in a later process.

 - Change sort key
Simple, change of sort key.
The sorted by name is sorted by priority of SORT, READ, name.