Japanese search
 - Create an index using Lucine
 - Assuming that data-struct, jp-sorting is resolved

The version of Lucine has changed dramatically, and since the Analyzer
for Japanese is used, architects and APIs are quite different.
The deviation between Airsonic and the source code is great.

The details of Analize may be different from Airsonic.
Adjust at any time, consider merging when Airsonic's Lucine is upgraded.