Refactoring SearchService
Divide SearchService into functions.
This Refactoring solves some of the individual problems at the same
time, but may create new inconsistencies.
(This point does not merge to the master)

This risk is caused by the lack of valid test cases for SearchService.

The goal of Refactoring is to eliminate the risk ultimately by properly
dividing the functions and adding test cases to the functions.

 - can not specify year #161
 Resolved below
 QueryFactory#createQuery(RandomSearchCriteria criteria)

 - Delete Lock operation #160
 Resolved below
 SearchService

  - Specification change of stopward #151
 Resolved below
 AnalyzerFactory

  - Update to Lucene 7.7.1 #121
  - CVE-2019-9658 #167
pom.xml