Security: Explicitly configure no-op password encoder
Spring Security 5 requires[1] storing the encoder id alongside the encoded password
(e.g. "{md5}hash" for an MD5-encoded password hash), which differs from previous
versions.

Airsonic unfortunately stores passwords in plain-text, which is why we are setting
the "no-op" (plain-text) password encoder as a default here. This default will be
used when no encoder id is present.

This means that legacy Airsonic passwords (stored simply as "password" in the db)
will be matched like "{noop}password" and will be recognized successfully. In the
future password encoding updates will be done here.

This fixes the following tests failures:

    [ERROR] Tests run: 1, Failures: 0, Errors: 1, Skipped: 0, Time elapsed: 9.656 s <<< FAILURE! - in org.airsonic.player.api.AirsonicRestApiIntTest
    [ERROR] pingTest  Time elapsed: 0.376 s  <<< ERROR!
    java.lang.IllegalArgumentException: There is no PasswordEncoder mapped for the id "null"
            at org.airsonic.player.api.AirsonicRestApiIntTest.pingTest(AirsonicRestApiIntTest.java:47)

[1] https://spring.io/blog/2017/11/01/spring-security-5-0-0-rc1-released#password-storage-updated