Apply format with Eclipse (#836)
 - Organize import
 - Source format
 - mvn formatter: format

The diff that appear in this commit are the blank line before the
package statement, the import statement, and the blank line at the end
of the file.
These cannot be applied using the current Maven formatter, indicating
that some automation relies on the IDE formatter.
It is a check target of Google check style.

Signed-off-by: tesshucom <webmaster@tesshu.com>