Margin setting at the bottom of the element (#887)
 - Fix bug where the bottom of the mainframe scrollbar was obscured by the player
 - Fix bug where the bottom of the mainframe content was obscured by the player

These are phenomena that have been confirmed mainly on some mobile
devices. The behavior depends on the browser, but it simplifies the
implementation so that it does not become a matrix. It sets margins by
default and cancels out extra margins if the agent is Windows. (In an
unknown environment, a large margin is set)