Add an option to search by analysis value instead of sort tag
In Jpsonic, parsed values exist for most tags, but these are not used
for searching. This option uses the parsed value instead of the tag for
the search. This feature is used when Japanese and alphabets are mixed
in the library, but you want to input search only in alphabets.