Disable some tests on Windows Server
Temporarily disabled due to some tests not working.
It has become unstable since a certain time, and it does not work even
if it is checked with a revision that was working fine.
t issue, not reproducible in production.

Operating System
Microsoft Windows Server 2022
10.0.20348
Datacenter

Runner Image
Image: windows-2022
Version: 20221127.1 (frequently)
Version: 20221204.3 (terrible)