Remove unnecessary dependencies
This is related to CVE-2025-48924. Unlike in the past, Commons Lang is
no longer used anywhere in this project.
Currently, only Commons Lang 3 (lang3) is implicitly referenced, and it
is up to date.
Therefore, the project is not affected by the CVE warning.

However, the outdated dependency on commons-lang is still declared in
the project’s POM, which may trigger a false positive in Docker Hub's
security scans.To prevent such false alerts, the unnecessary dependency
should be removed.