Add CacheDisposer, SmartLifecycleLegacyDaoHelper
The scan shutdown process is migrated from
the legacy implementation to SmartLifecycle.

Previously, it was implemented in a J2EE-compliant manner,
but that approach lacked the capabilities
to ensure precise shutdown ordering.
Therefore, relying on Spring’s mechanisms is preferred.

HSQLDB shutdown previously depended on @PreDestroy,
but it has been modified to be handled
via SmartLifecycleLegacyDaoHelper,
which implements SmartLifecycle.