Remove @DependsOn
The shutdown process for scan completion
is being migrated from
the legacy implementation to SmartLifecycle.

Previously, it was implemented according to J2EE standards,
but that approach lacked the functionality
to guarantee a precise shutdown order.
Relying on Spring’s mechanisms is therefore preferable.

The SmartLifecycle shutdown process executes first,
after which each executor is safely destroyed.
Consequently, all @PreDestroy annotations
on executor definitions are removed.